pyproject.toml
[tool.poetry]
name = "photometadata"
version = "0.1.0"
description = "Scripts to verify photo metadata"
authors = ["James Robinson <james.em.robinson@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
cleo = "^0.8.1"
emoji = "^1.2.0"
ExifRead = "^2.3.2"
pendulum = "^2.1.2"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
